requirements.txt
discord-webhook
python-dotenv~=0.19.0
requests~=2.26.0